Code page tests
